chore: split out request-response data helpers
