--- src/views/open-api-docs-view/utils/get-body-parameter-props.ts
@@ -1,43 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import { PropertyDetailsProps } from '../components/property-details'
-import { getPropertyDetailPropsFromSchemaObject } from './'
-// Types
-import type { OpenAPIV3 } from 'openapi-types'
-
-/**
- * Given the parameter object from a body parameter,
- * which we expect to be an object with a schema if it exists,
- *
- * Return property detail data.
- */
-export async function getBodyParameterProps(
-	bodyParam: OpenAPIV3.ParameterObject
-): Promise<PropertyDetailsProps[]> {
-	// Skip the bodyParam.schema if it's a reference
-	// We don't expect references, but for typing purposes we handle them.
-	if ('$ref' in bodyParam.schema) {
-		return []
-	}
-	// We flatten these properties to avoid showing a
-	// "collapsed object" UI under the "Body Parameters" section,
-	// which would be a bit redundant and annoying to have to expand
-	const bodyProps = []
-	for (const [key, value] of Object.entries(bodyParam.schema.properties)) {
-		// Skip reference objects, we expect these to be de-referenced
-		// before this function is called
-		if ('$ref' in value) {
-			continue
-		}
-		// Skip read-only properties
-		if (value.readOnly) {
-			continue
-		}
-		// Push props
-		bodyProps.push(await getPropertyDetailPropsFromSchemaObject(key, value))
-	}
-	return bodyProps
-}
